refactor: Convert everything to ESM and increment minimum Node version (#69)
More and more of the ecosystem is moving ESM support for modules, and
with good reason. The point is that it's possible to import both ESM and
CommonJS modules from an ESM application, while a CommonJS application
can ony ever import CommonJS. ka-mensa-api so far has been built
around CommonJS. This means we were limited in the types of libraries
we could use. This is no longer the case. :)

Loading of the ka-mensa-fetch data files (canteens.json, legend.json)
provided a bit of a hurdle, but was easily solved with manually creating
a require() function.