chore: Ensure tests are also typechecked on lint (#117)
ESLint does not typecheck, which is why we are running 'tsc --noEmit' as
part of the 'lint' script. Unfortunately, this included only the 'src/'
directory up until now due to a missing '-p' (or '--project') flag.
This didn't cause any issues because type errors would also be reported
during the 'test' script since ts-node also does typechecking, but we
really want these errors to be found during linting.