fix: Implement custom `formatDate()` and `parseDate()` for speed
By avoiding the use of `moment`, we achieve the following performance
gains:

* `formatDate()`: approx. 58x speedup
* `parseDate()`: approx. 22x speedup

We also lose semantic validation of the date in `parseDate()`, i.e., the
function will no longer reject inputs such as `'2023-02-31'`. This isn't
relevant in our case, though, since it will still lead to a 404 response
at a later point.