test: Add integration tests for the HTTP routes (#223)
The recent migration from Express to Fastify caused some bugs in the
HTTP server, such as a wrong response format. This could've been caught
with integration tests. To prevent such problems in the future, this
patch adds tests for all routes to assert correct responses for valid
and also for invalid input. Implementing this required moving the HTTP
server into a separate module that's free of side effects.